feat : 비회원 처리를 위한 어노테이션 추가

+++ src/main/java/nextstep/subway/path/domain/DiscountPolicy.java
@@ -23,7 +23,7 @@ public enum DiscountPolicy {
             .orElse(ADULT);
     }
 
-    public double reduce(int fare) {
+    public double reduce(double fare) {
         return fare * rate;
     }
